% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weave.R
\name{argb32_set_dim}
\alias{argb32_set_dim}
\title{Set the image dimensions of the data in an argb32 object}
\usage{
argb32_set_dim(argb32_ptr, width, height)
}
\arguments{
\item{argb32_ptr}{argb32 pointer}

\item{width, height}{image dimensions in pixels}
}
\description{
Warning: if you set dimensions which don't match the data you'll
likely get a segfault.
}
