% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weave.R
\name{argb32_to_planar}
\alias{argb32_to_planar}
\title{Convert an object with packed ARGB32 pixels into array of planar colour or matrix}
\usage{
argb32_to_planar(argb32_ptr, format = "rgba")
}
\arguments{
\item{argb32_ptr}{An object containing ARGB32 data. this is an external pointer
to data.}

\item{format}{one of 'rgba', 'rgb', or 'grey' to indicate the result should
be an RGBA array, an RGB array or a matrix, respectively}
}
\description{
Convert an object with packed ARGB32 pixels into array of planar colour or matrix
}
