% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-conversion.R
\name{argb32_to_raw}
\alias{argb32_to_raw}
\title{Convert an ARGB32 pointer to a raw vector}
\usage{
argb32_to_raw(argb32_ptr)
}
\arguments{
\item{argb32_ptr}{ARGB32 pointer}
}
\value{
raw vector
}
\description{
This function copies the data in the ARGB32 pointer to a new raw vector and
returns the new raw vector
}
