% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{packed_fmt}
\alias{packed_fmt}
\title{Supported packed color formats}
\format{
An object of class \code{list} of length 3.
}
\usage{
packed_fmt
}
\description{
Supported packed color formats
}
\keyword{datasets}
