% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weave.R
\name{packed_ptr_set_dim}
\alias{packed_ptr_set_dim}
\title{Set the image dimensions of the data in an packed color object}
\usage{
packed_ptr_set_dim(packed_ptr, width, height)
}
\arguments{
\item{packed_ptr}{External pointer to data which contains packed color values}

\item{width, height}{image dimensions in pixels}
}
\description{
This function should be used if the External Pointer to data was created in
another library. In this case there would be no dimension information stored
within the object, and it would need to be set explicitly.
}
\details{
Warning: Be careful when using this function - if you set dimensions which
don't match the data you'll likely get a segfault.
}
