% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weave.R
\name{packed_set_dim}
\alias{packed_set_dim}
\title{Set the image dimensions of the data in an packed pixel object}
\usage{
packed_set_dim(packed_ptr, width, height)
}
\arguments{
\item{packed_ptr}{packed pixel pointer}

\item{width, height}{image dimensions in pixels}
}
\description{
Warning: if you set dimensions which don't match the data you'll
likely get a segfault.
}
