% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-conversion.R
\name{packed_to_raw}
\alias{packed_to_raw}
\title{Convert a packed data pointer to a raw vector}
\usage{
packed_to_raw(packed_ptr)
}
\arguments{
\item{packed_ptr}{pointer to data containing packed colors}
}
\value{
raw vector
}
\description{
This function copies the data in the packed pointer to a new raw vector and
returns the new raw vector
}
