% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weave.R
\name{planar_to_argb32}
\alias{planar_to_argb32}
\title{Convert a matrix or array into packed ARGB32 format}
\usage{
planar_to_argb32(robj, maxval)
}
\arguments{
\item{robj}{R representation. Either a matrix (representing grey values), or
an array with 3 or 4 planes (representing RGB and RGBA colour
respectively).  Data must be double preceision floating point.}

\item{maxval}{specify the maximum possible value in the \code{robj} data.
This is used to scale the results into the desired range [0, 255].
For standard R image representations, data is often in the range
[0,1] so use \code{maxval=1}}
}
\description{
Convert a matrix or array into packed ARGB32 format
}
