% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-conversion.R
\name{raw_to_argb32}
\alias{raw_to_argb32}
\title{Convert a raw vector into a ARGB32 pointer}
\usage{
raw_to_argb32(raw_vec, width, height)
}
\arguments{
\item{raw_vec}{raw vector}

\item{width, height}{image dimensions}
}
\value{
An external poitner to ARGB32 data
}
\description{
This function copies the original data into a new object and returns a
pointer to this data.
}
