% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle packed colors in-place}
\usage{
shuffle(packed_raw, src_fmt, dst_fmt)
}
\arguments{
\item{packed_raw}{Packed color stored in a vector of raw values}

\item{src_fmt, dst_fmt}{Format specification. See \code{pixelweaver::packed_fmt}}
}
\value{
Nothing.  The vector is shuffled in-place.
}
\description{
Shuffle packed colors in-place
}
