% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-pam.R
\name{write_pam}
\alias{write_pam}
\title{Write array to a NETPBM PAM file}
\usage{
write_pam(arr, filename, maxval = 1)
}
\arguments{
\item{arr}{matrix (grey) or array (with rgb or rgba data)}

\item{filename}{output filename e.g. \code{"output.pam"}}

\item{maxval}{maximum value possible in the array representation. For R
arrays this is often 1.}
}
\description{
PAM file is written with depth 4 with \code{TUPLTYPE = RGB_ALPHA}
}
