% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-pam.R
\name{write_pam_raw}
\alias{write_pam_raw}
\title{Write packed color in ABGR32 format to a NETPBM PAM file}
\usage{
write_pam_raw(packed_raw, filename, width, height)
}
\arguments{
\item{packed_raw}{Packed color data in ABGR32 format}

\item{filename}{output filename e.g. \code{"output.pam"}}

\item{width, height}{image dimensions}
}
\description{
PAM file is written with depth 4 with \code{TUPLTYPE = RGB_ALPHA}
}
