% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weave.R
\name{packed_to_planar}
\alias{packed_to_planar}
\title{Convert an object with packed pixels into array of color values in R}
\usage{
packed_to_planar(packed_ptr, format, nchannel)
}
\arguments{
\item{packed_ptr}{An object containing packed pixel data. This is an external pointer
to data.}

\item{format}{Integer value denoting the packed format. 0L = ARGB32,
1L = RGBA32, 2L = ABGR32.  See also \code{?pack_fmt}.}

\item{nchannel}{Number of color channels in the planar representation. 1 = grey,
3 = RGB, 4 = RGBA}
}
\description{
Convert an object with packed pixels into array of color values in R
}
